requirements.txt
# Betty Framework Requirements

# Core dependencies
PyYAML>=6.0.1
pyyaml>=6.0.1

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
mypy>=1.5.0
types-PyYAML>=6.0.12
